New: retrieve whole directory

+++ FileRetriever.py
@@ -1,6 +1,15 @@
 from ftplib import FTP
 import os
 
+def retrieveFolderContent(directory):
+    # create a ftp connection with the Wilber ftp server
+    ftp = FTP("ds.iris.edu")
+    ftp.login() 
+
+    # go to the provided directory
+    ftp.cwd("pub/userdata/wilber/" + directory)
+    return ftp.nlst()
+
 # retrieve the data from Wilber 3: http://ds.iris.edu/wilber3
 def retrieveWilberData(directory, outputDirectory):
     # create output directory if it doesn't exist yet
@@ -19,9 +28,14 @@ def retrieveWilberData(directory, outputDirectory):
     # download all files
     for filename in filenames:
         local_filename = os.path.join(os.getcwd() + "/" + outputDirectory, filename)
-        print("Downloading " + filename)
-        file = open(local_filename, "wb")
-        ftp.retrbinary("RETR "+ filename, file.write)
-        file.close()
+
+        # only download the non existing files
+        if not os.path.exists(local_filename):
+            print("Downloading " + filename)
+            file = open(local_filename, "wb")
+            ftp.retrbinary("RETR "+ filename, file.write)
+            file.close()
+        else:
+            print("Skipping " + filename)
 
     ftp.quit()
